deepen: fold cached-summary + fallback-notice into First Onboarding Email
Consolidate `lib/onboarding/cached-summary-delivery.ts` and
`lib/email/onboarding-fallback-service.ts` into one deep module,
`lib/onboarding/first-email-delivery.ts`, behind a single
`sendFirstOnboardingEmail(...)` interface. The module orchestrates the
cached-summary and fallback-notice paths behind one seam so callers
never branch on which path fires.

Both prior modules independently checked and wrote
`User.onboardingFirstEmailSentAt`, so the idempotency invariant lived
in two places. The consolidated module has one guard and one atomic
marker write for both paths — locality gained. The cached path
additionally writes `User.onboardingFirstSummaryId` and a
`SummaryEmailDelivery` row; the fallback path does not. Result reports
`path: 'cached' | 'fallback'` so callers distinguish outcomes without
re-computing.

Callers updated:
- `app/(auth)/onboarding/actions.ts` after() block — was two sequential
  dynamic imports with an explicit `if (reason === no_cached_summaries
  || no_tickers)` fallback dispatch. Now one call.
- `app/api/onboarding?action=deliver-summaries` — used the deprecated
  `deliverCachedSummaries` shim (cached-only). Now calls the
  consolidated function, which falls back to notice if no cache. This
  is a small correctness improvement: the endpoint previously reported
  `{ delivered: 0, reason: 'no_cached_summaries' }` for the long-tail
  ticker case; it now delivers a fallback notice. In practice the
  action.ts after() block runs first and sets the idempotency marker,
  so this branch fires only for pre-marker callers.

Tests replaced: `cached-summary-delivery.test.ts` (516 LOC, direct
unit tests of `calculateCompositeScore` and the two internal delivery
functions) and `onboarding-fallback-service.test.ts` (211 LOC) →
`first-email-delivery.test.ts` (21 tests at the deepened module's
interface). The internal scoring, candidate selection, tiebreaker, and
template lookup are exercised only through observable outcomes at
`sendFirstOnboardingEmail`. Same pattern as ADR-0002.

Dependency category: **local-substitutable** (Prisma + Resend mocked in
tests, per [process/Deepening/deepening.md](process/Deepening/deepening.md)
§2). No new port introduced — one adapter (production) is a hypothetical
seam.

CONTEXT.md updated with a new **First Onboarding Email** term
cross-referencing [Onboarding], [Summary], and the same-shape
inlined-guard cluster (Why It Matters Guard, Ticker Grounding Guard,
etc.).
