+++ CONTEXT.md
@@ -24,6 +24,9 @@ A user's saved interest in a [Filing] type or ticker. Drives which [Summary] pay
 ### Onboarding
 The first-run flow that turns a new signup into an engaged subscriber: ticker selection, preview email, account creation, first cached [Summary] delivery.
 
+### First Onboarding Email
+The single email a new signup receives immediately after completing [Onboarding], as the payoff for the "you'll receive an email shortly" promise on the onboarding action. Implemented as one deep module (`lib/onboarding/first-email-delivery.ts`) behind a single `sendFirstOnboardingEmail({ userId, userEmail, recipientName, trackedTickers })` function. The module orchestrates two internal delivery paths behind one seam: the **cached path** picks the strongest cached [Summary] across the user's tickers by a composite score (importance × 0.25 + materiality × 0.40 + analysisDepth × 0.20 + recency × 0.15, capped at 50 candidates) and dispatches through `sendFilingSummaryEmail`; when no candidates exist (long-tail unique-ticker or zero-ticker case), the **fallback path** sends the "we're watching your tickers" notice via `EmailType.ONBOARDING_FALLBACK_NOTICE`. Both paths share one idempotency guard on `User.onboardingFirstEmailSentAt` and one atomic marker write, so re-invocation is a no-op regardless of which path fires. The cached path additionally writes `User.onboardingFirstSummaryId` and a `SummaryEmailDelivery` row tagged `source: onboarding-best-pick`; the fallback path writes neither (no [Summary] was sent). Result reports `path: 'cached' | 'fallback'` so callers can distinguish outcomes without re-computing. Replaces two drifting modules — `lib/onboarding/cached-summary-delivery.ts` (cached path + a `deliverCachedSummaries` deprecated shim) and `lib/email/onboarding-fallback-service.ts` (fallback path) — that each carried their own copy of the idempotency guard AND their own write of `onboardingFirstEmailSentAt`, so a change to the idempotency invariant had to be made in two places. Consuming callers (`app/(auth)/onboarding/actions.ts` after() block and `/api/onboarding?action=deliver-summaries`) now call the one function; the cached-then-fallback branching lives inside the module. Sub-scoring (analysis depth, composite score, tiebreaker) is private implementation and only exercised through the module's interface — see `__tests__/lib/onboarding/first-email-delivery.test.ts`. Same shape as [Email Link Token] / [Template Selection] / [Tier Scheduling].
+
 ### X Sentiment
 An enrichment of a [Summary] that captures public discussion on X (Twitter) about the ticker over a 7-day window. Direction (bullish/bearish/mixed/neutral/no_signal), shift, confidence, and citation-backed claims. Gated by form importance + a large-cap allowlist (S&P 500) to keep cost and pump-and-dump risk bounded. Runs only for paid-tier callers via the enrichment gate.
