+++ app/(auth)/onboarding/actions.ts
@@ -458,49 +458,33 @@ export async function completeOnboardingBatched(input: {
     // User.onboardingFirstEmailSentAt.
     after(async () => {
       try {
-        const { deliverFirstOnboardingEmail } = await import(
-          '@/lib/onboarding/cached-summary-delivery'
+        const { sendFirstOnboardingEmail } = await import(
+          '@/lib/onboarding/first-email-delivery'
         );
-        const deliveryResult = await deliverFirstOnboardingEmail(
-          result.id,
-          primaryEmail
-        );
+        const deliveryResult = await sendFirstOnboardingEmail({
+          userId: result.id,
+          userEmail: primaryEmail,
+          recipientName: userName,
+          trackedTickers: input.tickers.map((t) => t.symbol),
+        });
         console.log(
-          '[Onboarding] Cached-summary delivery:',
+          '[Onboarding] First-email delivery:',
           JSON.stringify({
             userId: result.id,
             delivered: deliveryResult.delivered,
+            path: deliveryResult.path,
             reason: deliveryResult.reason,
             summaryId: deliveryResult.summaryId,
             score: deliveryResult.score,
           })
         );
 
-        // Long-tail fallback: no cached summaries available across the user's
-        // tickers (unique-ticker case). Send the dedicated fallback notice so
-        // the "you'll receive an email shortly" promise isn't broken silently.
-        if (
-          !deliveryResult.delivered &&
-          (deliveryResult.reason === 'no_cached_summaries' ||
-            deliveryResult.reason === 'no_tickers')
-        ) {
-          const { sendOnboardingFallbackNotice } = await import(
-            '@/lib/email/onboarding-fallback-service'
-          );
-          await sendOnboardingFallbackNotice({
-            userId: result.id,
-            email: primaryEmail,
-            recipientName: userName,
-            trackedTickers: input.tickers.map((t) => t.symbol),
-          });
-        }
-
-        // TODO: if reason === 'no_cached_summaries', enqueue
+        // TODO: if path === 'fallback', enqueue
         // ASYNC_DISCOVER_FILINGS for the user's tickers to shorten the wait
         // for their first real filing email. Requires understanding cron auth
         // + idempotency-key pattern. Deferred to follow-up — see TODOS.md.
       } catch (err) {
-        console.error('[Onboarding] Cached-summary delivery threw:', err);
+        console.error('[Onboarding] First-email delivery threw:', err);
       }
     });
